:art: Rename flow/milestone.hpp to flow/step.hpp
A step is a milestone or an action, and the header is for both.
