+++ include/flow/flow.hpp
@@ -5,5 +5,5 @@
 #include <flow/detail/par.hpp>
 #include <flow/detail/seq.hpp>
 #include <flow/impl.hpp>
-#include <flow/milestone.hpp>
 #include <flow/run.hpp>
+#include <flow/step.hpp>
