+++ include/flow/graph_builder.hpp
@@ -3,7 +3,6 @@
 #include <flow/common.hpp>
 #include <flow/detail/walk.hpp>
 #include <flow/impl.hpp>
-#include <flow/milestone.hpp>
 
 #include <stdx/ct_string.hpp>
 #include <stdx/cx_multimap.hpp>
